fix: move show config error alert to components
