+++ src/app/core/services/config.service.ts
@@ -19,9 +19,6 @@ import { FirebaseAnalyticsService } from './firebaseAnalytics.service'
 import { NotificationService } from './notification.service'
 import { SchedulingService } from './scheduling.service'
 import { StorageService } from './storage.service'
-import { AlertController } from 'ionic-angular'
-import { TranslatePipe } from '../../shared/pipes/translate/translate'
-import { LocKeys } from '../../shared/enums/localisations'
 @Injectable()
 export class ConfigService {
   constructor(
@@ -30,9 +27,7 @@ export class ConfigService {
     private schedule: SchedulingService,
     private notificationService: NotificationService,
     private appVersion: AppVersion,
-    private firebaseAnalytics: FirebaseAnalyticsService,
-    private alertCtrl: AlertController,
-    private translate: TranslatePipe
+    private firebaseAnalytics: FirebaseAnalyticsService
   ) {}
 
   fetchConfigState(force: boolean) {
@@ -43,137 +38,100 @@ export class ConfigService {
       this.appVersion.getVersionNumber()
     ]).then(
       ([configVersion, scheduleVersion, storedAppVersion, appVersion]) => {
-        return this.pullProtocol()
-          .then(res => {
-            if (res) {
-              const response: any = JSON.parse(res)
-              if (
-                configVersion !== response.version ||
-                scheduleVersion !== response.version ||
-                storedAppVersion !== appVersion ||
-                force
-              ) {
-                this.storage.set(StorageKeys.APP_VERSION, appVersion)
-                this.storage.set(StorageKeys.HAS_CLINICAL_TASKS, false)
-                const protocolFormated = this.formatPulledProcotol(
-                  response.protocols
-                )
-                const scheduledAssessments = []
-                const clinicalAssessments = []
-                for (let i = 0; i < protocolFormated.length; i++) {
-                  const clinical =
-                    protocolFormated[i]['protocol']['clinicalProtocol']
-                  if (clinical) {
-                    this.storage.set(StorageKeys.HAS_CLINICAL_TASKS, true)
-                    clinicalAssessments.push(protocolFormated[i])
-                  } else {
-                    scheduledAssessments.push(protocolFormated[i])
-                  }
+        return this.pullProtocol().then(res => {
+          if (res) {
+            const response: any = JSON.parse(res)
+            if (
+              configVersion !== response.version ||
+              scheduleVersion !== response.version ||
+              storedAppVersion !== appVersion ||
+              force
+            ) {
+              this.storage.set(StorageKeys.APP_VERSION, appVersion)
+              this.storage.set(StorageKeys.HAS_CLINICAL_TASKS, false)
+              const protocolFormated = this.formatPulledProcotol(
+                response.protocols
+              )
+              const scheduledAssessments = []
+              const clinicalAssessments = []
+              for (let i = 0; i < protocolFormated.length; i++) {
+                const clinical =
+                  protocolFormated[i]['protocol']['clinicalProtocol']
+                if (clinical) {
+                  this.storage.set(StorageKeys.HAS_CLINICAL_TASKS, true)
+                  clinicalAssessments.push(protocolFormated[i])
+                } else {
+                  scheduledAssessments.push(protocolFormated[i])
                 }
-                return this.storage
-                  .set(StorageKeys.CONFIG_VERSION, response.version)
-                  .then(() => {
-                    return this.storage
-                      .set(
-                        StorageKeys.CONFIG_CLINICAL_ASSESSMENTS,
-                        clinicalAssessments
-                      )
-                      .then(() => {
-                        console.log('Pulled clinical questionnaire')
-                        return this.pullQuestionnaires(
-                          StorageKeys.CONFIG_CLINICAL_ASSESSMENTS
-                        )
-                      })
-                  })
-                  .then(() => {
-                    return this.storage
-                      .set(StorageKeys.CONFIG_ASSESSMENTS, scheduledAssessments)
-                      .then(() => {
-                        console.log('Pulled questionnaire')
-                        return this.pullQuestionnaires(
-                          StorageKeys.CONFIG_ASSESSMENTS
-                        )
-                      })
-                  })
-                  .then(() => this.schedule.generateSchedule(true))
-                  .then(() => this.rescheduleNotifications())
-                  .then(() =>
-                    Promise.all([
-                      this.storage.get(StorageKeys.PARTICIPANTLOGIN),
-                      this.getProjectName(),
-                      this.storage.get(StorageKeys.SOURCEID),
-                      this.storage.get(StorageKeys.ENROLMENTDATE),
-                      this.storage.get(StorageKeys.PARTICIPANTID)
-                    ]).then(
-                      ([
-                        subjectId,
-                        projectId,
-                        sourceId,
-                        enrolmentDate,
-                        humanReadableId
-                      ]) =>
-                        this.firebaseAnalytics.setUserProperties({
-                          subjectId: subjectId,
-                          projectId: projectId,
-                          sourceId: sourceId,
-                          enrolmentDate: String(enrolmentDate),
-                          humanReadableId: humanReadableId
-                        })
-                    )
-                  )
-                  .then(() =>
-                    this.firebaseAnalytics.logEvent('config_update', {
-                      config_version: String(configVersion),
-                      schedule_version: String(scheduleVersion),
-                      app_version: appVersion
-                    })
-                  )
-              } else {
-                console.log(
-                  'NO CONFIG UPDATE. Version of protocol.json has not changed.'
-                )
-                return this.schedule.generateSchedule(false)
               }
+              return this.storage
+                .set(StorageKeys.CONFIG_VERSION, response.version)
+                .then(() => {
+                  return this.storage
+                    .set(
+                      StorageKeys.CONFIG_CLINICAL_ASSESSMENTS,
+                      clinicalAssessments
+                    )
+                    .then(() => {
+                      console.log('Pulled clinical questionnaire')
+                      return this.pullQuestionnaires(
+                        StorageKeys.CONFIG_CLINICAL_ASSESSMENTS
+                      )
+                    })
+                })
+                .then(() => {
+                  return this.storage
+                    .set(StorageKeys.CONFIG_ASSESSMENTS, scheduledAssessments)
+                    .then(() => {
+                      console.log('Pulled questionnaire')
+                      return this.pullQuestionnaires(
+                        StorageKeys.CONFIG_ASSESSMENTS
+                      )
+                    })
+                })
+                .then(() => this.schedule.generateSchedule(true))
+                .then(() => this.rescheduleNotifications())
+                .then(() =>
+                  Promise.all([
+                    this.storage.get(StorageKeys.PARTICIPANTLOGIN),
+                    this.getProjectName(),
+                    this.storage.get(StorageKeys.SOURCEID),
+                    this.storage.get(StorageKeys.ENROLMENTDATE),
+                    this.storage.get(StorageKeys.PARTICIPANTID)
+                  ]).then(
+                    ([
+                      subjectId,
+                      projectId,
+                      sourceId,
+                      enrolmentDate,
+                      humanReadableId
+                    ]) =>
+                      this.firebaseAnalytics.setUserProperties({
+                        subjectId: subjectId,
+                        projectId: projectId,
+                        sourceId: sourceId,
+                        enrolmentDate: String(enrolmentDate),
+                        humanReadableId: humanReadableId
+                      })
+                  )
+                )
+                .then(() =>
+                  this.firebaseAnalytics.logEvent('config_update', {
+                    config_version: String(configVersion),
+                    schedule_version: String(scheduleVersion),
+                    app_version: appVersion
+                  })
+                )
+            } else {
+              console.log(
+                'NO CONFIG UPDATE. Version of protocol.json has not changed.'
+              )
+              return this.schedule.generateSchedule(false)
             }
-          })
-          .catch(e => {
-            console.log(e)
-            return this.showConfigError()
-          })
+          }
+        })
       }
     )
-  }
-
-  showConfigError() {
-    const buttons = [
-      {
-        text: this.translate.transform(LocKeys.BTN_OKAY.toString()),
-        handler: () => {
-          this.fetchConfigState(true)
-        }
-      }
-    ]
-    this.showAlert({
-      title: this.translate.transform(LocKeys.STATUS_FAILURE.toString()),
-      message: this.translate.transform(LocKeys.PROTOCOL_ERROR_DESC.toString()),
-      buttons: buttons
-    })
-  }
-
-  showAlert(parameters) {
-    const alert = this.alertCtrl.create({
-      title: parameters.title,
-      buttons: parameters.buttons
-    })
-    if (parameters.message) {
-      alert.setMessage(parameters.message)
-    }
-    if (parameters.inputs) {
-      for (let i = 0; i < parameters.inputs.length; i++) {
-        alert.addInput(parameters.inputs[i])
-      }
-    }
-    alert.present()
   }
 
   rescheduleNotifications() {
