+++ src/app/pages/settings/containers/settings-page.component.ts
@@ -279,10 +279,27 @@ export class SettingsPageComponent {
     this.showLoading = true
     return this.configService
       .fetchConfigState(true)
+      .catch(e => this.showConfigError())
       .then(() => {
         this.showLoading = false
         return this.loadSettings()
       })
       .then(() => this.backToSplash())
+  }
+
+  showConfigError() {
+    const buttons = [
+      {
+        text: this.translate.transform(LocKeys.BTN_OKAY.toString()),
+        handler: () => {
+          this.reloadConfig()
+        }
+      }
+    ]
+    this.showAlert({
+      title: this.translate.transform(LocKeys.STATUS_FAILURE.toString()),
+      message: this.translate.transform(LocKeys.PROTOCOL_ERROR_DESC.toString()),
+      buttons: buttons
+    })
   }
 }
